0804, 6 Dec 19, build complete

+++ HW3/WinButton.cpp
@@ -18,7 +18,7 @@ namespace HW3
 		if (width > this->name.length() + 4)
 		{
 			width -= 4;
-			width -= this->name.length();
+			width -= static_cast<int>(this->name.length());
 		}
 		else
 		{
@@ -36,7 +36,7 @@ namespace HW3
 	int WinButton::drawState(std::ostream& os) const
 	{
 		int width = getWidth();
-		width -= getStateStr().length() + 2;
+		width -= static_cast<int>(getStateStr().length()) + 2;
 		os << "|";
 		while (width > 0)
 		{
@@ -50,7 +50,7 @@ namespace HW3
 	int WinButton::drawValue(std::ostream& os) const
 	{
 		int width = getWidth();
-		width -= this->value.length() + 2;
+		width -= static_cast<int>(this->value.length()) + 2;
 		os << "|";
 		while (width > 0)
 		{
